Put future sub indices in one file
