+++ corehq/apps/es/const.py
@@ -37,7 +37,7 @@ HQ_CASE_SEARCH_INDEX_CANONICAL_NAME = "case_search"
 HQ_CASE_SEARCH_INDEX_NAME = "case-search-20230524"
 HQ_CASE_SEARCH_SECONDARY_INDEX_NAME = "case-search-2024-05-09"
 
-# See case_search_bha.py docstring for context
+# See case_search_sub.py docstring for context
 HQ_CASE_SEARCH_BHA_INDEX_CANONICAL_NAME = "case_search_bha"
 HQ_CASE_SEARCH_BHA_INDEX_NAME = "case-search-bha-2024-05-10"
 HQ_CASE_SEARCH_BHA_SECONDARY_INDEX_NAME = None
